ath6kl: handle concurrent AP-STA channel switches
If an ath6kl AP vif is beaconing on one channel, and a STA vif
associates on a different channel, a WMI_DISCONNECT event will be sent
to the AP vif. Make the AP vif follow the STA interface, and notify
userspace.

kvalo: fix a sparse warning with vif->next_chan

Signed-off-by: Thomas Pedersen <c_tpeder@qca.qualcomm.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>

Conflicts:

	drivers/net/wireless/ath/ath6kl/cfg80211.c
	drivers/net/wireless/ath/ath6kl/cfg80211.h
	drivers/net/wireless/ath/ath6kl/core.h